# docs/swagger.yaml
basePath: /api/v1
definitions:
  main.AddToCartPayload:
    properties:
      product_id:
        type: integer
      quantity:
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
...
  main.CreateUserTokenPayload:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 72
        minLength: 3
        type: string
    required:
    - email
    - password
    type: object
  main.RegisterUserPayload:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 72
        minLength: 3
        type: string
      username:
        maxLength: 100
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  store.Cart:
    properties:
      created_at:
        type: string
      id:
        type: integer
      stores:
        description: Relasi
        items:
          $ref: '#/definitions/store.CartStores'
        type: array
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  store.CartItem:
    properties:
      cart_id:
        type: integer
      cart_store_id:
        description: Nullable di database
        type: string
      created_at:
        type: string
      id:
        type: string
      product:
        allOf:
        - $ref: '#/definitions/store.Product'
        description: Relasi
      product_id:
        type: integer
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  store.CartStores:
    properties:
      cart_id:
        type: integer
      created_at:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/store.CartItem'
        type: array
      toko:
        allOf:
        - $ref: '#/definitions/store.Toko'
        description: Relasi
      toko_id:
        type: integer
    type: object
  store.Category:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
...
  store.Product:
    properties:
      category:
        $ref: '#/definitions/store.Category'
      country:
        type: string
      created_at:
        type: string
      description:
        type: string
      discount:
        type: number
      discount_price:
        type: number
      estimation:
        type: string
      id:
        type: integer
      image_urls:
        items:
          type: string
        type: array
      is_approved:
        type: boolean
      is_for_sale:
        type: boolean
      name:
        type: string
      price:
        type: number
      rating:
        type: number
      slug:
        type: string
      sold:
        type: integer
      stock:
        type: integer
      toko:
        $ref: '#/definitions/store.Toko'
      updated_at:
        type: string
      version:
        type: integer
    type: object
...
  store.SingleUser:
    properties:
      created_at:
        type: string
      email:
        type: string
      google_id:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      picture:
        type: string
      username:
        type: string
    type: object
  store.Toko:
    properties:
      country:
        type: string
      created_at:
        type: string
      id:
        type: integer
      image_profile:
        type: string
      name:
        type: string
      slug:
        type: string
      user:
        $ref: '#/definitions/store.SingleUser'
      user_id:
        type: integer
    type: object
...
info:
  contact:
    email: mochammad.yogaprasetya112@gmail.com
    name: API Support
    url: http://www.myogaprasetya.my.id
  description: |-
    API ini menggunakan mekanisme rate limiter untuk membatasi jumlah permintaan yang dapat dilakukan oleh setiap klien dalam jangka waktu tertentu.
    Hal ini bertujuan untuk mencegah penyalahgunaan API, memastikan kinerja yang stabil, dan melindungi server dari potensi serangan seperti brute force atau DDoS.
    Rate limiter ini didukung oleh Redis sebagai penyimpanan sementara yang cepat dan andal untuk melacak jumlah permintaan setiap klien berdasarkan API key atau alamat IP mereka.
    Redis digunakan karena kemampuannya yang tinggi dalam menangani data secara real-time dengan latensi rendah.
    Jika batas permintaan terlampaui, klien akan menerima respons dengan kode status 429 (Too Many Requests).
  termsOfService: http://www.myogaprasetya.my.id
  title: Tokopedia API
paths:
  /authentication/logout:
    get:
      description: Logs out a user
      produces:
      - application/json
      responses:
        "204":
          description: User logged out
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema: {}
      summary: Logs out a user
      tags:
      - authentication
  /authentication/token:
    get:
      consumes:
      - application/json
      description: Creates a token for a user
      parameters:
      - description: User credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CreateUserTokenPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Token
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Creates a token
      tags:
      - authentication
  /authentication/user:
    post:
      consumes:
      - application/json
      description: Registers a user
      parameters:
      - description: User credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.RegisterUserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: User registered
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Registers a user
      tags:
      - authentication
  /cart:
    get:
      consumes:
      - application/json
      description: fetch a cart
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      - description: sort
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Cart'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: fetch a cart
      tags:
      - cart
    post:
      consumes:
      - application/json
      description: create a cart
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.AddToCartPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Cart'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: create a cart
      tags:
      - cart
  /cart/item/{cartStoreItemID}:
    get:
      consumes:
      - application/json
      description: fetch a cart store item by ID
      parameters:
      - description: cart store item ID
        in: path
        name: cartStoreItemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Cart'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: fetch a cart store item
      tags:
      - cart
  /cart/item/{cartStoreItemID}/decrease:
    patch:
      consumes:
      - application/json
      description: remove quantity to cart store item
      parameters:
      - description: cart store item ID
        in: path
        name: cartStoreItemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Cart'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: remove quantity to cart store item
      tags:
      - cart
  /cart/item/{cartStoreItemID}/delete:
    delete:
      consumes:
      - application/json
      description: delete cart store item by ID
      parameters:
      - description: cart store item ID
        in: path
        name: cartStoreItemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: delete cart store item
      tags:
      - cart
  /cart/item/{cartStoreItemID}/increase:
    patch:
      consumes:
      - application/json
      description: add quantity to cart store item
      parameters:
      - description: cart store item ID
        in: path
        name: cartStoreItemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Cart'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: add quantity to cart store item
      tags:
      - cart
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"